# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 착한가격업소 찾기 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Store {
  id: number;
  /**
   * 원본 데이터의 번호
   * @nullable
   */
  externalId: number | null;
  /** 업종명 */
  category: string;
  /** 업소명 */
  name: string;
  /** 주요품목 */
  mainItem: string;
  /** 가격(원) */
  price: number;
  /**
   * 업소 전화번호
   * @nullable
   */
  phone: string | null;
  /** 주소 */
  address: string;
  /** 위도 */
  latitude: number;
  /** 경도 */
  longitude: number;
  /**
   * 네이버지도 URL
   * @nullable
   */
  naverMapUrl: string | null;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface StoresStats {
  total: number;
  averagePrice: number;
  categoryBreakdown: CategoryCount[];
  /**
   * ISO 8601 timestamp
   * @nullable
   */
  lastUpdated: string | null;
}

export interface StoreInput {
  /** @nullable */
  externalId?: number | null;
  category: string;
  name: string;
  mainItem: string;
  price: number;
  /** @nullable */
  phone?: string | null;
  address: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  naverMapUrl?: string | null;
}

export interface ImportStoresResponse {
  inserted: number;
  updated: number;
  total: number;
}

export interface ErrorResponse {
  error: string;
}

export type SuggestionStatus =
  (typeof SuggestionStatus)[keyof typeof SuggestionStatus];

export const SuggestionStatus = {
  pending: "pending",
  confirmed: "confirmed",
} as const;

export interface Suggestion {
  id: number;
  storeId: number;
  /**
   * 업소명 (조회시에만 포함)
   * @nullable
   */
  storeName: string | null;
  content: string;
  status: SuggestionStatus;
  /** ISO 8601 timestamp */
  createdAt: string;
}

export type DeleteStore200 = {
  deleted: number;
};

export type ImportStoresBody = {
  stores: StoreInput[];
};

export type ListSuggestionsParams = {
  status?: ListSuggestionsStatus;
};

export type ListSuggestionsStatus =
  (typeof ListSuggestionsStatus)[keyof typeof ListSuggestionsStatus];

export const ListSuggestionsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  all: "all",
} as const;

export type CreateSuggestionBody = {
  storeId: number;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  content: string;
};

export type UpdateSuggestionBodyStatus =
  (typeof UpdateSuggestionBodyStatus)[keyof typeof UpdateSuggestionBodyStatus];

export const UpdateSuggestionBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
} as const;

export type UpdateSuggestionBody = {
  status: UpdateSuggestionBodyStatus;
};

export type DeleteSuggestion200 = {
  deleted: number;
};