# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 착한가격업소 찾기 API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stores
    description: 착한가격업소 데이터
  - name: suggestions
    description: 정보 수정 제안
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: 전체 착한가격업소 목록 조회
      description: 위도/경도/업종/가격 등 정렬 및 필터링은 클라이언트에서 처리
      responses:
        "200":
          description: 업소 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
  /stores/stats:
    get:
      operationId: getStoresStats
      tags: [stores]
      summary: 업소 통계 정보
      description: 등록된 업소 수, 업종 분포, 평균 가격
      responses:
        "200":
          description: 통계 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoresStats"
  /stores/{id}:
    patch:
      operationId: updateStore
      tags: [stores]
      summary: 업소 정보 수정 (관리자)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "200":
          description: 수정된 업소
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          description: 잘못된 입력
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 업소를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: 업소 삭제 (관리자)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 삭제 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                required:
                  - deleted
        "404":
          description: 업소를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/import:
    post:
      operationId: importStores
      tags: [stores]
      summary: 엑셀 데이터로 업소 병합 (관리자)
      description: 네이버지도 URL을 기준으로 기존 데이터와 병합 (있으면 update, 없으면 insert)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stores:
                  type: array
                  items:
                    $ref: "#/components/schemas/StoreInput"
              required:
                - stores
      responses:
        "200":
          description: 병합 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportStoresResponse"
        "400":
          description: 잘못된 입력
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /suggestions:
    get:
      operationId: listSuggestions
      tags: [suggestions]
      summary: 정보 수정 제안 목록 (관리자)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, confirmed, all]
      responses:
        "200":
          description: 제안 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Suggestion"
    post:
      operationId: createSuggestion
      tags: [suggestions]
      summary: 정보 수정 제안 등록 (이용자)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storeId:
                  type: integer
                content:
                  type: string
                  minLength: 1
                  maxLength: 2000
              required:
                - storeId
                - content
      responses:
        "200":
          description: 등록 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Suggestion"
        "400":
          description: 잘못된 입력
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 업소를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /suggestions/{id}:
    patch:
      operationId: updateSuggestion
      tags: [suggestions]
      summary: 제안 상태 변경 (관리자)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, confirmed]
              required:
                - status
      responses:
        "200":
          description: 변경된 제안
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Suggestion"
        "404":
          description: 제안을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSuggestion
      tags: [suggestions]
      summary: 제안 삭제 (관리자)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 삭제 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                required:
                  - deleted
        "404":
          description: 제안을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Store:
      type: object
      properties:
        id:
          type: integer
        externalId:
          type: ["integer", "null"]
          description: 원본 데이터의 번호
        category:
          type: string
          description: 업종명
        name:
          type: string
          description: 업소명
        mainItem:
          type: string
          description: 주요품목
        price:
          type: integer
          description: 가격(원)
        phone:
          type: ["string", "null"]
          description: 업소 전화번호
        address:
          type: string
          description: 주소
        latitude:
          type: number
          description: 위도
        longitude:
          type: number
          description: 경도
        naverMapUrl:
          type: ["string", "null"]
          description: 네이버지도 URL
      required:
        - id
        - externalId
        - category
        - name
        - mainItem
        - price
        - phone
        - address
        - latitude
        - longitude
        - naverMapUrl
    StoresStats:
      type: object
      properties:
        total:
          type: integer
        averagePrice:
          type: integer
        categoryBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        lastUpdated:
          type: ["string", "null"]
          description: ISO 8601 timestamp
      required:
        - total
        - averagePrice
        - categoryBreakdown
        - lastUpdated
    CategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required:
        - category
        - count
    StoreInput:
      type: object
      properties:
        externalId:
          type: ["integer", "null"]
        category:
          type: string
        name:
          type: string
        mainItem:
          type: string
        price:
          type: integer
        phone:
          type: ["string", "null"]
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        naverMapUrl:
          type: ["string", "null"]
      required:
        - category
        - name
        - mainItem
        - price
        - address
        - latitude
        - longitude
    ImportStoresResponse:
      type: object
      properties:
        inserted:
          type: integer
        updated:
          type: integer
        total:
          type: integer
      required:
        - inserted
        - updated
        - total
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Suggestion:
      type: object
      properties:
        id:
          type: integer
        storeId:
          type: integer
        storeName:
          type: ["string", "null"]
          description: 업소명 (조회시에만 포함)
        content:
          type: string
        status:
          type: string
          enum: [pending, confirmed]
        createdAt:
          type: string
          description: ISO 8601 timestamp
      required:
        - id
        - storeId
        - storeName
        - content
        - status
        - createdAt