# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 착한가격업소 찾기 API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * 위도/경도/업종/가격 등 정렬 및 필터링은 클라이언트에서 처리
 * @summary 전체 착한가격업소 목록 조회
 */
export const ListStoresResponseItem = zod.object({
  id: zod.number(),
  externalId: zod.number().nullable().describe("원본 데이터의 번호"),
  category: zod.string().describe("업종명"),
  name: zod.string().describe("업소명"),
  mainItem: zod.string().describe("주요품목"),
  price: zod.number().describe("가격(원)"),
  phone: zod.string().nullable().describe("업소 전화번호"),
  address: zod.string().describe("주소"),
  latitude: zod.number().describe("위도"),
  longitude: zod.number().describe("경도"),
  naverMapUrl: zod.string().nullable().describe("네이버지도 URL"),
});
export const ListStoresResponse = zod.array(ListStoresResponseItem);

/**
 * 등록된 업소 수, 업종 분포, 평균 가격
 * @summary 업소 통계 정보
 */
export const GetStoresStatsResponse = zod.object({
  total: zod.number(),
  averagePrice: zod.number(),
  categoryBreakdown: zod.array(
    zod.object({
      category: zod.string(),
      count: zod.number(),
    }),
  ),
  lastUpdated: zod.string().nullable().describe("ISO 8601 timestamp"),
});

/**
 * @summary 업소 정보 수정 (관리자)
 */
export const UpdateStoreParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateStoreBody = zod.object({
  externalId: zod.number().nullish(),
  category: zod.string(),
  name: zod.string(),
  mainItem: zod.string(),
  price: zod.number(),
  phone: zod.string().nullish(),
  address: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  naverMapUrl: zod.string().nullish(),
});

export const UpdateStoreResponse = zod.object({
  id: zod.number(),
  externalId: zod.number().nullable().describe("원본 데이터의 번호"),
  category: zod.string().describe("업종명"),
  name: zod.string().describe("업소명"),
  mainItem: zod.string().describe("주요품목"),
  price: zod.number().describe("가격(원)"),
  phone: zod.string().nullable().describe("업소 전화번호"),
  address: zod.string().describe("주소"),
  latitude: zod.number().describe("위도"),
  longitude: zod.number().describe("경도"),
  naverMapUrl: zod.string().nullable().describe("네이버지도 URL"),
});

/**
 * @summary 업소 삭제 (관리자)
 */
export const DeleteStoreParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteStoreResponse = zod.object({
  deleted: zod.number(),
});

/**
 * 네이버지도 URL을 기준으로 기존 데이터와 병합 (있으면 update, 없으면 insert)
 * @summary 엑셀 데이터로 업소 병합 (관리자)
 */
export const ImportStoresBody = zod.object({
  stores: zod.array(
    zod.object({
      externalId: zod.number().nullish(),
      category: zod.string(),
      name: zod.string(),
      mainItem: zod.string(),
      price: zod.number(),
      phone: zod.string().nullish(),
      address: zod.string(),
      latitude: zod.number(),
      longitude: zod.number(),
      naverMapUrl: zod.string().nullish(),
    }),
  ),
});

export const ImportStoresResponse = zod.object({
  inserted: zod.number(),
  updated: zod.number(),
  total: zod.number(),
});

/**
 * @summary 정보 수정 제안 목록 (관리자)
 */
export const ListSuggestionsQueryParams = zod.object({
  status: zod.enum(["pending", "confirmed", "all"]).optional(),
});

export const ListSuggestionsResponseItem = zod.object({
  id: zod.number(),
  storeId: zod.number(),
  storeName: zod.string().nullable().describe("업소명 (조회시에만 포함)"),
  content: zod.string(),
  status: zod.enum(["pending", "confirmed"]),
  createdAt: zod.string().describe("ISO 8601 timestamp"),
});
export const ListSuggestionsResponse = zod.array(ListSuggestionsResponseItem);

/**
 * @summary 정보 수정 제안 등록 (이용자)
 */
export const createSuggestionBodyContentMax = 2000;

export const CreateSuggestionBody = zod.object({
  storeId: zod.number(),
  content: zod.string().min(1).max(createSuggestionBodyContentMax),
});

export const CreateSuggestionResponse = zod.object({
  id: zod.number(),
  storeId: zod.number(),
  storeName: zod.string().nullable().describe("업소명 (조회시에만 포함)"),
  content: zod.string(),
  status: zod.enum(["pending", "confirmed"]),
  createdAt: zod.string().describe("ISO 8601 timestamp"),
});

/**
 * @summary 제안 상태 변경 (관리자)
 */
export const UpdateSuggestionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSuggestionBody = zod.object({
  status: zod.enum(["pending", "confirmed"]),
});

export const UpdateSuggestionResponse = zod.object({
  id: zod.number(),
  storeId: zod.number(),
  storeName: zod.string().nullable().describe("업소명 (조회시에만 포함)"),
  content: zod.string(),
  status: zod.enum(["pending", "confirmed"]),
  createdAt: zod.string().describe("ISO 8601 timestamp"),
});

/**
 * @summary 제안 삭제 (관리자)
 */
export const DeleteSuggestionParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteSuggestionResponse = zod.object({
  deleted: zod.number(),
});