# lib/api-zod/src/generated/types/suggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 착한가격업소 찾기 API
 * OpenAPI spec version: 0.1.0
 */
import type { SuggestionStatus } from "./suggestionStatus";

export interface Suggestion {
  id: number;
  storeId: number;
  /**
   * 업소명 (조회시에만 포함)
   * @nullable
   */
  storeName: string | null;
  content: string;
  status: SuggestionStatus;
  /** ISO 8601 timestamp */
  createdAt: string;
}